+++ src/client/components/user-preview.vue
@@ -33,7 +33,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import parseAcct from '@/misc/acct/parse';
+import { parseAcct } from '@/misc/acct';
 import MkFollowButton from './follow-button.vue';
 import { userPage } from '../filters/user';
 import * as os from '@client/os';
