+++ src/misc/acct.ts
@@ -1,0 +1,14 @@
+export type Acct = {
+	username: string;
+	host: string | null;
+};
+
+export const getAcct = (user: Acct) => {
+	return user.host == null ? user.username : `${user.username}@${user.host}`;
+};
+
+export const parseAcct = (acct: string): Acct => {
+	if (acct.startsWith('@')) acct = acct.substr(1);
+	const split = acct.split('@', 2);
+	return { username: split[0], host: split[1] || null };
+};
