--- src/misc/acct/render.ts
@@ -1,5 +1,0 @@
-import Acct from './type';
-
-export default (user: Acct) => {
-	return user.host == null ? user.username : `${user.username}@${user.host}`;
-};
