--- src/misc/acct/type.ts
@@ -1,6 +1,0 @@
-type Acct = {
-	username: string;
-	host: string | null;
-};
-
-export default Acct;
