+++ src/server/api/endpoints/pinned-users.ts
@@ -1,7 +1,7 @@
 import define from '../define';
 import { Users } from '../../../models';
 import { fetchMeta } from '@/misc/fetch-meta';
-import parseAcct from '@/misc/acct/parse';
+import { parseAcct } from '@/misc/acct';
 import { User } from '../../../models/entities/user';
 
 export const meta = {
