+++ src/server/api/limiter.ts
@@ -1,7 +1,7 @@
 import * as Limiter from 'ratelimiter';
 import { redisClient } from '../../db/redis';
 import { IEndpoint } from './endpoints';
-import getAcct from '@/misc/acct/render';
+import { getAcct } from '@/misc/acct';
 import { User } from '../../models/entities/user';
 import Logger from '../../services/logger';
