+++ src/server/well-known.ts
@@ -1,8 +1,7 @@
 import * as Router from '@koa/router';
 
 import config from '@/config';
-import parseAcct from '@/misc/acct/parse';
-import Acct from '@/misc/acct/type';
+import { parseAcct, Acct } from '@/misc/acct';
 import { links } from './nodeinfo';
 import { escapeAttribute, escapeValue } from '../prelude/xml';
 import { Users } from '../models';
